+++ CHANGES.md
@@ -180,6 +180,8 @@ TODO
    :class:`mangadap.util.sampling.Resample`.
  - Add a class that reads trace sets produced by IDLUTILS,
    :class:`mangadap.util.trace.TraceSet`, but not well tested yet.
+ - Fix untested velocity registration and stacking code, and added
+   registration test
  - Upgraded package requirements
  - Line profile in `EmissionLineTemplates` now hardwired to be
    `FFTGaussianLSF`.